==> 02_Code/05_NAN_Analysis.ipynb <==
# importing required libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import warnings

# Setting default options
warnings.filterwarnings("ignore")
pd.set_option("mode.copy_on_write", True)

# loading the Data
df = pd.read_csv("/Users/revanth/Downloads/insurance_data_cleaned (1).csv")
df.sample(5)

df.describe()

df.info()

# check if any actual null values
df.isnull().sum()

Column by column checks

df["YEAR"].value_counts().sort_index()

df["SEX"].value_counts()

df["RACENEW"].value_counts().sort_index()

consider 530 and 541 as the NaN values

df["INCFAM97ON2"].value_counts().sort_index()

0 as Nan, should we consider 96 as well?

df["HINOTCOVE"].value_counts().sort_index()

df["CNLUNG"].value_counts().sort_index()

df["SMK"].value_counts().sort_index()

df["Occupation_Code"].value_counts().sort_index()

## Analysis

Only The following columns are required for the analysis:
- Race
- Income
- Smoke
- Occupation

# creating a subset of the data with only the columns we need
required_columns = ["RACENEW", "INCFAM97ON2", "SMK", "Occupation_Code"]

df_analysis = df[required_columns]

# rename the columns
col_map = {
    "RACENEW": "race",
    "INCFAM97ON2": "inc",
    "SMK": "smk",
    "Occupation_Code": "occ",
}
df_analysis.rename(columns=col_map, inplace=True)
df_analysis.sample(5)

# Remapping rules,
# Add the values which need to be considered as 1 in remapping, rest will be considered as 0
remap_race = [530, 541]
remap_smk = [0]
remap_inc = [0]
remap_occ = [0]

def converter(val, remap_list):
    # takes in a value and  a list of values to consider as "missing" which will be marked as 1 and the rest will be marked as 0
    if val in remap_list:
        return 1
    else:
        return 0

# applying remapping rules
df_analysis["inc"] = df_analysis["inc"].apply(lambda x: converter(x, remap_inc))
df_analysis["occ"] = df_analysis["occ"].apply(lambda x: converter(x, remap_occ))
df_analysis["race"] = df_analysis["race"].apply(lambda x: converter(x, remap_race))
df_analysis["smk"] = df_analysis["smk"].apply(lambda x: converter(x, remap_smk))

def any_null(row):
    # check if any value in the row is 1 and returns 1 if any value is 1
    if 1 in row.values:
        return 1
    else:
        return 0

# applying funtion to the dataframe
df_analysis["any_null"] = df_analysis.apply(lambda x: any_null(x), axis=1)

df_analysis.sample(5)

df_analysis.describe()

# function to exclude columns and calculate the any_nul
def any_null_exclude(row, exclude_lst):
    # check if any value in the row is 1 and returns 1 if any value is 1
    if 1 in row.drop(exclude_lst).values:
        return 1
    else:
        return 0

# excluding the occ and any_null columns
df_analysis["any_null_exclude_occ"] = df_analysis.apply(
    lambda x: any_null_exclude(x, ["occ", "any_null"]), axis=1
)

df_analysis.describe()

# excluding the inc and any_null columns
df_analysis["any_null_exclude_occ_inc"] = df_analysis.apply(
    lambda x: any_null_exclude(x, ["occ", "any_null", "inc", "any_null_exclude_occ"]),
    axis=1,
)

df_analysis.describe()
