==> 02_Code/08_No_NIU.ipynb <==
# importing required libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import warnings

# Setting default options
warnings.filterwarnings("ignore")
pd.set_option("mode.copy_on_write", True)

# loading the Data
# df = pd.read_csv("../01_Data/01_Raw/insurance_data_cleaned.csv")
df = pd.read_csv("/Users/revanth/Downloads/insurance_data_cleaned (1).csv")
df.sample(5)

df.info()

df.describe()

rows_before = df.shape[0]

# NIU List, values which need to be removed
remap_race = [530, 541]
remap_smk = [0]
remap_inc = [0]
remap_occ = [0]

# Removing the NIU values
df = df[~df["RACENEW"].isin(remap_race)]
df = df[~df["SMK"].isin(remap_smk)]
df = df[~df["INCFAM97ON2"].isin(remap_inc)]
df = df[~df["Occupation_Code"].isin(remap_occ)]

rows_after = df.shape[0]

print(f"Inital Rows: {rows_before}")
print(f"Final Rows: {rows_after}")
print(f"Rows Removed: {rows_before - rows_after}")
print(f"Percentage Removed: {((rows_before - rows_after) / rows_before) * 100:.2f}%")

Validation

df_cols = df.columns

for col in df_cols:
    print(f"{col}: {df[col].unique()}")

# saving the cleaned data
df.to_csv("../01_Data/02_Processed/01_Insurance_NoMissing.csv", index=False)
